+++ src/cli/upload/stdin-to-uss-file/StdinToUssFile.Handler.ts
@@ -11,7 +11,7 @@
 
 import { FTPBaseHandler } from "../../../FTPBase.Handler";
 import { IFTPHandlerParams } from "../../../IFTPHandlerParams";
-import { CoreUtils, UssUtils, TRANSFER_TYPE_ASCII, TRANSFER_TYPE_BINARY } from "../../../api";
+import { CoreUtils, UssUtils } from "../../../api";
 
 export default class UploadStdinToUssFileHandler extends FTPBaseHandler {
 
@@ -23,7 +23,7 @@ export default class UploadStdinToUssFileHandler extends FTPBaseHandler {
 
         const options = {
             content,
-            transferType: params.arguments.binary ? TRANSFER_TYPE_BINARY : TRANSFER_TYPE_ASCII,
+            transferType: CoreUtils.getBinaryTransferModeOrDefault(params.arguments.binary),
         };
         await UssUtils.uploadFile(params.connection, ussFile, options);